More Information
I wrote this script for an industry project to go through and automatically test the accuracy of several hundred random forest models and output the results of these accuracy tests to a file. For this project, we focused on comparing precision, recall, cumulative lift, accuracy, kappa, specificity, and sensitivity. Since there were so many datasets being modeled and tested, I utilized both the makePSOCKcluster function and an on-campus supercomputer cluster to run these analyses in parallel.